Modularize types
