+++ types/Link.d.ts
@@ -1,0 +1,24 @@
+import { SvelteComponentTyped } from 'svelte'
+import { RouteLocation } from './Route'
+
+interface LinkProps {
+  to: string
+  replace?: boolean
+  state?: {
+    [k in string | number]: unknown
+  }
+  getProps?: (linkParams: GetPropsParams) => Record<string, any>
+}
+
+interface GetPropsParams {
+  location: RouteLocation
+  href: string
+  isPartiallyCurrent: boolean
+  isCurrent: boolean
+}
+
+type Link = SvelteComponentTyped<
+  Omit<LinkProps & svelte.JSX.HTMLProps<HTMLAnchorElement> & svelte.JSX.SapperAnchorProps, 'href'>
+>
+
+export default Link
