+++ types/Router.d.ts
@@ -1,0 +1,10 @@
+import { SvelteComponentTyped } from 'svelte'
+
+interface RouterProps {
+  basepath?: string
+  url?: string
+}
+
+type Router = SvelteComponentTyped<RouterProps>
+
+export default Router
