+++ types/actions.d.ts
@@ -1,0 +1,4 @@
+type link = (node: Element) => { destroy(): void }
+type links = (node: Element) => { destroy(): void }
+
+export { link, links }
