+++ types/functions.d.ts
@@ -1,0 +1,14 @@
+type navigate = (
+  to: string,
+  {
+    replace,
+    state
+  }: {
+    replace?: boolean
+    state: {
+      [k in string | number]: unknown
+    }
+  }
+) => void
+
+export { navigate }
